refactor: move module overlap detection code to a separate file
